refactor: rename ModalBackground to ModalBackground.styles
