+++ src/components/ModalBackground.styles.tsx
