+++ src/components/__tests__/ModalBackground.styles.test.tsx
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react';
-import ModalBackground from '../ModalBackground';
+import ModalBackground from '../ModalBackground.styles';
 
 describe('<ModalBackground />', (): void => {
   it('renders the component successfully', (): void => {
@@ -13,16 +13,8 @@ describe('<ModalBackground />', (): void => {
     const element = getByTestId(testId);
     expect(element.tagName.toLowerCase()).toEqual('div');
     expect(element).toHaveStyle(`
-      position: fixed;
-      top: 0;
-      left: 0;
       width: 100%;
       height: 100%;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      background-color: rgba(0,0,0,0.5);
-      z-index: 30;
     `);
   });
 });
